services/place/requirements.txt
alembic==0.8.10
coverage==4.5.3
Flask==1.0.2
coverage==4.5.3
flake8===3.7.7
flask-migrate==2.4.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
flask-debugtoolbar==0.10.1
Flask-Testing==0.7.1
psycopg2-binary==2.7.7
flask-cors==3.0.7